feat: Create ICCAs build scripts & exclude ICCA screens from main build
Use Metro build config to conditionally include ICCA screens so that
ICCA resouces are not included in the main Mapeo build. This avoids the
images used in the ICCAs intro screens increasing the file size of the
main Mapeo APK.
