+++ android/app/src/icca/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+  <string name="app_name">Mapeo for ICCAs</string>
+</resources>
