+++ android/app/src/iccaDebug/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+  <string name="app_name">Mapeo ICCAs Dev</string>
+</resources>
