+++ package.json
@@ -7,13 +7,15 @@
   },
   "scripts": {
     "start": "react-native start",
+    "start-icca": "RN_SRC_EXT=icca.js react-native start",
     "android": "npm run build:backend && npm run android-no-backend-rebuild",
     "android-no-backend-rebuild": "react-native run-android --variant=appDebug --appIdSuffix=debug",
     "android-storybook": "react-native run-android --variant=storybookDebug --appIdSuffix=storybook.debug",
-    "android-icca": "react-native run-android --variant=iccaDebug --appIdSuffix=debug",
+    "android-icca": "RN_SRC_EXT=icca.js react-native run-android --variant=iccaDebug --appIdSuffix=icca.debug",
     "build:backend": "./scripts/build-backend.sh",
     "build:translations": "node ./scripts/build-translations.js",
     "build:release": "npm run build:translations && npm run build:backend && ./scripts/build-release-android.sh",
+    "build:release-icca": "npm run build:translations && npm run build:backend && RN_SRC_EXT=icca.js ./scripts/build-release-android.sh",
     "build:storybook": "./node_modules/@storybook/react/bin/build.js -c ./storybook-web -s ./public -o ./build --no-dll",
     "test": "jest",
     "lint": "eslint *.js \"src/**/*.js\"",
