+++ src/frontend/AppContainer.icca.js
@@ -1,0 +1,17 @@
+// @flow
+import AppStack from "./AppStack";
+import IntroStack from "./screens/Intro";
+import { createAppContainer, createSwitchNavigator } from "react-navigation";
+
+const RootStack = createSwitchNavigator(
+  {
+    Intro: IntroStack,
+    App: AppStack,
+  },
+  {
+    initialRouteName: "Intro",
+  }
+);
+
+// $FlowFixMe
+export default createAppContainer(RootStack);
